add FullInvoice type to supplier invoice business
